:recycle: refacto and moved a bunch of code for the sake of readability
