+++ imageCompressor.cabal
@@ -25,6 +25,8 @@ source-repository head
 
 library
   exposed-modules:
+      CentroidBind
+      DisplayResult
       FileParsing
       ImageCompressor
       Parsing
