+++ src/CentroidBind.hs
@@ -1,0 +1,37 @@
+{-
+-- EPITECH PROJECT, 2024
+-- ImageCompressor
+-- File description:
+-- CentroidBind
+-}
+
+module CentroidBind (CentroidBind(..),
+                     createRandomCentroids,
+                     bindVectorToCentroid,
+                     emptyCentroid) where
+
+import Vector (Vector(..), getClosestVector)
+
+data CentroidBind = CentroidBind Vector [(Vector, Vector)] deriving (Show)
+
+createRandomCentroids :: Int -> [CentroidBind]
+createRandomCentroids 0 = []
+createRandomCentroids n =
+  CentroidBind (Vector [128, 128, 128]) [] : createRandomCentroids (n - 1)
+
+getCentroidFromBindList :: [CentroidBind] -> [Vector]
+getCentroidFromBindList [] = []
+getCentroidFromBindList ((CentroidBind centroid _):xs) =
+  centroid : getCentroidFromBindList xs
+
+bindVectorToCentroid :: (Vector, Vector) -> [CentroidBind] -> [CentroidBind]
+bindVectorToCentroid _ [] = []
+bindVectorToCentroid (coord, vec) ((CentroidBind centroid lPair):xs)
+  | centroid == getClosestVector vec
+    (getCentroidFromBindList ((CentroidBind centroid lPair):xs)) =
+      (CentroidBind centroid ((coord, vec):lPair)) : xs
+  | otherwise =
+    (CentroidBind centroid lPair) : bindVectorToCentroid (coord, vec) xs
+
+emptyCentroid :: CentroidBind -> CentroidBind
+emptyCentroid (CentroidBind centroid _) = CentroidBind centroid []
