+++ src/Vector.hs
@@ -2,10 +2,12 @@
 -- EPITECH PROJECT, 2024
 -- ImageCompressor
 -- File description:
--- Parsing
+-- Vector
 -}
 
-module Vector (Vector(..), getDistance) where
+module Vector (Vector(..), getDistance, getClosestVector, roundVector) where
+
+import GHC.Float
 
 data Vector = Vector [Double] deriving (Show)
 
@@ -16,3 +18,14 @@ instance Eq Vector where
 getDistance :: Vector -> Vector -> Double
 getDistance (Vector l1) (Vector l2) =
   sqrt (sum (zipWith (\a b -> (a - b) ** 2) l1 l2))
+
+getClosestVector :: Vector -> [Vector] -> Vector
+getClosestVector _ [] = error "Empty list"
+getClosestVector _ [a] = a
+getClosestVector v (a:xs)
+  | getDistance v a < getDistance v (getClosestVector v xs) = a
+  | otherwise = getClosestVector v xs
+
+roundVector :: Vector -> Vector
+roundVector (Vector l) =
+  Vector (map (\x -> int2Double (double2Int (x + 0.5))) l)
